make console panel compile correctly in vc7.1, wx2.6
git-svn-id: svn://svn.code.sf.net/p/magicseteditor/code/trunk@1632 0fc631ac-6414-0410-93d0-97cfa31319b6

+++ src/gui/set/console_panel.cpp
@@ -99,7 +99,7 @@ class MessageCtrl : public wxScrolledWindow {
 		Refresh(false);
 	}
 	
-	int find_point(int y) {
+	size_t find_point(int y) {
 		// TODO: could do a binary search here
 		for (size_t i = 0 ; i < messages.size() ; ++i) {
 			if (y >= messages[i]->top && y < messages[i]->bottom()) return i;
